Do not use rpm --specfile to query the subpackage names
Using "rpm --specfile..." is "insecure", because during macro
expansion arbitrary code can be executed ("%(...)"). Instead,
use the helpers/spec_query script, which does not expand such
macros, to list the names of all subpackages. Hence, this might
break some packages.

Note: we also change the semantics, because we ignore all %if*
conditionals during the specfile parsing. That is, we also
take subpackages into account that are guarded by some %if...
The advantage of this change is that we get the same results
on all architectures.

Fixes: boo#938556