Fix --keep-name-conditionals code in helpers/spec_query
The old code, which was introduced in commit e1a20d2 ("Add
--keep-name-conditionals option to helpers/spec_query"), was completely
wrong, because it just kept the "%if %name..." line, but not the
corresponding "%endif" line. Now, we treat name conditionals as follows:

%if %name...
<IFPART>
%endif

is rewritten to

%if %name...
<IFPART'>
%endif

where <IFPART'> is the result of applying the if-conditional handling
code to the <IFPART> block.
Moreover,

%if %name...
<IFPART>
%else
<ELSEPART>
%endif

is rewritten to

%if %name...
<IFPART'>
%endif
<ELSEPART'>

where <IFPART'> and <ELSEPART'> are the results of applying the
if-conditional handling code to <IFPART> and <ELSEPART>, respectively.

Fixes: #61 ('source_validator complains about "not mentioned" files
in spec although they are (coreutils)')